requirements.txt
streamlit
google-genai
pydantic
python-dotenv
edge-tts
pygame
moviepy<2.0.0
numpy
Pillow
replicate
demucs
openai-whisper
pydub
ffmpeg-python
yt-dlp
